+++ build-scripts/baptizer.sh
@@ -13,6 +13,7 @@ function usage() {
     echo -e "-l PLUGINS_HASH\t[Optional] Commit hash from the wazuh-indexer-plugins repository"
     echo -e "-e REPORTING_HASH\t[Optional] Commit hash from the wazuh-indexer-reporting repository"
     echo -e "-s SECURITY_HASH\t[Optional] Commit hash from the wazuh-indexer-security-analytics repository"
+    echo -e "-n NOTIFICATIONS_HASH\t[Optional] Commit hash from the wazuh-indexer-notifications repository"
     echo -e "-m MIN\t[Optional] Use naming convention for minimal packages, default is 'false'."
     echo -e "-x RELEASE\t[Optional] Use release naming convention, default is 'false'."
     echo -e "-h help"
@@ -23,7 +24,7 @@ function usage() {
 # ====
 function parse_args() {
 
-    while getopts ":hp:a:d:r:l:e:s:mx" arg; do
+    while getopts ":hp:a:d:r:l:e:s:n:mx" arg; do
         case $arg in
         h)
             usage
@@ -49,6 +50,9 @@ function parse_args() {
             ;;
         s)
             SECURITY_HASH=$OPTARG
+            ;;
+        n)
+            NOTIFICATIONS_HASH=$OPTARG
             ;;
         m)
             IS_MIN=true
@@ -141,8 +145,8 @@ function get_devel_name() {
         PREFIX="$PREFIX"-min
     fi
     # Generate composed commit hash
-    if [ -n "$PLUGINS_HASH" ] && [ -n "$REPORTING_HASH" ] && [ -n "$SECURITY_HASH" ]; then
-        COMMIT_HASH="$GIT_COMMIT"-"$PLUGINS_HASH"-"$REPORTING_HASH"-"$SECURITY_HASH"
+    if [ -n "$PLUGINS_HASH" ] && [ -n "$REPORTING_HASH" ] && [ -n "$SECURITY_HASH" ] && [ -n "$NOTIFICATIONS_HASH" ]; then
+        COMMIT_HASH="$GIT_COMMIT"-"$PLUGINS_HASH"-"$REPORTING_HASH"-"$SECURITY_HASH"-"$NOTIFICATIONS_HASH"
     fi
     PACKAGE_NAME="$PREFIX"_"$VERSION"-"$REVISION"_"$SUFFIX"_"$COMMIT_HASH"."$EXT"
 }